fix: convert ints to doubles feat: MessageUtil::format (configurable in config.yml) feat: new placeholders

+++ api/src/main/java/ovh/mythmc/banco/api/Banco.java
@@ -19,7 +19,7 @@ public interface Banco {
 
     boolean isOnline(UUID uuid);
 
-    int getInventoryValue(UUID uuid);
+    double getInventoryValue(UUID uuid);
 
     void clearInventory(UUID uuid);
